Change structure member type.
Former-commit-id: 41cab0d462702da128cb318f4e26e0e51b28891f

+++ client/src/include/scrollbar.h
@@ -62,7 +62,7 @@ typedef struct scrollbar_struct
 	/** Maximum number of lines. */
 	uint32 max_lines;
 
-	int *redraw;
+	uint8 *redraw;
 
 	SDL_Surface *surface;
